Added ascii image cog

+++ src/cogs/amogus.py
@@ -15,9 +15,22 @@ class Amogus(commands.Cog):
 
     @commands.command()
     async def amogus(self, ctx):
-        await ctx.message.delete()
-        await ctx.send(AMOGUS_SOUND_FILE)
-        await AudioManager.playAudio(ctx.voice_client, AMOGUS_SOUND_FILE, allowLocal=True)
+        voiceClient = ctx.voice_client
+        voiceClient.stop()  # In case it was already playing something
+
+        # Check the sender is in a voice channel
+        if ctx.author.voice is not None:
+            # Move to the senders voice channel
+            voiceChannel = ctx.author.voice.channel
+            if voiceClient is None:
+                await voiceChannel.connect();
+            else:
+                await voiceClient.move_to(voiceChannel)
+            # Play the sound
+            await AudioManager.playAudio(ctx.voice_client, AMOGUS_SOUND_FILE, allowLocal=True)
+            await ctx.message.delete()
+        else:
+            await ctx.send("You are not in a voice channel!")
 
 def setup(client):
     client.add_cog(Amogus(client))
